Add the additional field to the form
The current form is `vendor/spryker-shop/customer-page/src/SprykerShop/Yves/CustomerPage/Form/ProfileForm.php`

We need to adjust this form in order to display an additional field for the antelope.

Extend is the only way when it comes to forms. We created a new ProfileForm class that extends from the Spryker-class

The FIELD_ANTELOPE constant set the field name, in this case fk_antelope

The addAntelopeField method will add the new field and set the properties of it and set the addModelTransformer as well using the Factory to us the createAntelopeTransformer