Make the new form available
Only the base classes like Layer-)Factories, DependencyProviders, Facades, or Configs are auto-located. For all the others, you need to instantiate them somewhere

For forms, we use a proxy-factory, which is a real facotry with helpers of spryker factories but it will not be auto-loaded, because it is not the module factory.

We created a FormFactory that extend from vendor/spryker-shop/customer-page/src/SprykerShop/Yves/CustomerPage/Form/FormFactory to use the project level form

Then from the CustomerPageFactory we create a createCustomerFormFactory method that create a new instance of FormFactory

With this the form field should be visible but will not persist anything